move open_connection_to_shard_aware_port() to connection
