+++ scylla/src/connection/connection_pool.rs
@@ -27,6 +27,8 @@ use std::time::Duration;
 use tokio::sync::{broadcast, mpsc, Notify};
 use tracing::{debug, error, trace, warn};
 
+use super::connection::open_connection_to_shard_aware_port;
+
 /// The target size of a per-node connection pool.
 #[derive(Debug, Clone, Copy)]
 pub enum PoolSize {
@@ -1189,29 +1191,6 @@ struct OpenedConnectionEvent {
     result: Result<(Connection, ErrorReceiver), ConnectionError>,
     requested_shard: Option<Shard>,
     keyspace_name: Option<VerifiedKeyspaceName>,
-}
-
-async fn open_connection_to_shard_aware_port(
-    endpoint: UntranslatedEndpoint,
-    shard: Shard,
-    sharder: Sharder,
-    connection_config: &ConnectionConfig,
-) -> Result<(Connection, ErrorReceiver), ConnectionError> {
-    // Create iterator over all possible source ports for this shard
-    let source_port_iter = sharder.iter_source_ports_for_shard(shard);
-
-    for port in source_port_iter {
-        let connect_result =
-            connection::open_connection(endpoint.clone(), Some(port), connection_config).await;
-
-        match connect_result {
-            Err(err) if err.is_address_unavailable_for_use() => continue, // If we can't use this port, try the next one
-            result => return result,
-        }
-    }
-
-    // Tried all source ports for that shard, give up
-    Err(ConnectionError::NoSourcePortForShard(shard))
 }
 
 #[cfg(test)]
